moved some test from non-vrf to vrf folder
