+++ contracts/test/v0.1/vrf/CoordinatorRevertFulfillment.test.cjs
@@ -7,17 +7,21 @@ const {
   parseRandomWordsRequestedTx,
   fulfillRandomWords,
   parseRandomWordsFulfilledTx
-} = require('../vrf/VRFCoordinator.utils.cjs')
+} = require('./VRFCoordinator.utils.cjs')
 const { parseKlay } = require('../utils.cjs')
 const {
   deploy: deployRrCoordinator,
   setupOracle: setupRequestResponseCoordinator,
   parseDataRequestedTx,
   parseDataRequestFulfilledTx
-} = require('./RequestResponseCoordinator.utils.cjs')
-const { deploy: deployPrepayment, createAccount, deposit } = require('./Prepayment.utils.cjs')
-const { vrfConfig } = require('../vrf/VRFCoordinator.config.cjs')
-const { requestResponseConfig } = require('./RequestResponse.config.cjs')
+} = require('../non-vrf/RequestResponseCoordinator.utils.cjs')
+const {
+  deploy: deployPrepayment,
+  createAccount,
+  deposit
+} = require('../non-vrf/Prepayment.utils.cjs')
+const { vrfConfig } = require('./VRFCoordinator.config.cjs')
+const { requestResponseConfig } = require('../non-vrf/RequestResponse.config.cjs')
 const { getBalance, createSigners } = require('../utils.cjs')
 const oraklVrf = import('@bisonai/orakl-vrf')
